Rename Milestone.vue to Milestones.
